# src/api/swagger/swaggerJson.json
{
	"swagger": "2.0",
	"info": {
		"description": "This is the Swagger Documentation implementation of API's on NodeJS. You will find detailed documentation and model data about the REST API's. To check if an API is working , execute the API after entering the valid parameters",
		"version": "1.0.0",
		"title": "Boiler",
		"termsOfService": "",
		"contact": {
			"email": "api@boiler.com"
		}
	},
	"host": "127.0.0.1:4001",
	"basePath": "/api/v1",
...
	"schemes": [
		"http"
	],
	"paths": {
...
		"/auth/forgot-password": {
			"post": {
				"tags": [
					"Authentication"
				],
				"summary": "forgot password",
				"description": "forgot password",
				"operationId": "forgotPassword",
				"consumes": [
					"application/json"
				],
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"description": "forgotPassword",
						"required": true,
						"schema": {
							"$ref": "#/definitions/forgotPassword"
						}
					},
					{
						"in": "header",
						"name": "lang",
						"description": "User language",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "Successfully sent mail"
					},
					"401": {
						"description": "Invalid username"
					},
					"500": {
						"description": "Internal Server Error"
					}
				}
			}
		},
		"/auth/reset-password": {
			"post": {
				"tags": [
					"Authentication"
				],
				"summary": "reset password",
				"description": "reset password",
				"operationId": "resetPassword",
				"consumes": [
					"application/json"
				],
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"description": "resetPassword",
						"required": true,
						"schema": {
							"$ref": "#/definitions/resetPassword"
						}
					},
					{
						"in": "header",
						"name": "lang",
						"description": "User language",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "Successfully changed password"
					},
					"401": {
						"description": "Invalid username"
					},
					"500": {
						"description": "Internal Server Error"
					}
				}
			}
		},
...
		"/auth/change-password": {
			"put": {
				"tags": [
					"Authentication"
				],
				"summary": "change password",
				"description": "change password",
				"operationId": "changePassword",
				"consumes": [
					"application/json"
				],
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"description": "resetPassword",
						"required": true,
						"schema": {
							"$ref": "#/definitions/changePassword"
						}
					},
					{
						"in": "header",
						"name": "x-access-token",
						"description": "accessToken",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "Successfully changed password"
					},
					"401": {
						"description": "Invalid username"
					},
					"500": {
						"description": "Internal Server Error"
					}
				}
			}
		},
...
		"/auth/send-otp": {
			"post": {
				"tags": [
					"Authentication"
				],
				"summary": "send otp to phone number for verification",
				"description": "send otp to phone number for verification",
				"operationId": "sendOtp",
				"consumes": [
					"application/json"
				],
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"description": "sendOtp",
						"required": true,
						"schema": {
							"$ref": "#/definitions/sendOtp"
						}
					},
					{
						"in": "header",
						"name": "x-access-token",
						"description": "accessToken",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "Successfully sent mail"
					},
					"401": {
						"description": "Invalid username"
					},
					"500": {
						"description": "Internal Server Error"
					}
				}
			}
		},
		"/auth/verify-phone": {
			"put": {
				"tags": [
					"Authentication"
				],
				"summary": "verify user phone number",
				"description": "verify user phone by otp",
				"operationId": "verifyPhone",
				"consumes": [
					"application/json"
				],
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"description": "verify user phone number by otp",
						"required": true,
						"schema": {
							"$ref": "#/definitions/verifyPhone"
						}
					},
					{
						"in": "header",
						"name": "x-access-token",
						"description": "accessToken",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "Email verified successfully"
					},
					"409": {
						"description": "Invalid input Error"
					},
					"500": {
						"description": "Internal Server Error"
					}
				}
			}
		},
...
	},
	"definitions": {
...
		"forgotPassword": {
			"type": "object",
			"required": [
				"email"
			],
			"properties": {
				"email": {
					"type": "string",
					"example": "john@blockgemini.com"
				}
			},
			"xml": {
				"name": "forgotPassword"
			}
		},
		"resetPassword": {
			"type": "object",
			"required": [
				"password",
				"code"
			],
			"properties": {
				"password": {
					"type": "string",
					"example": "password"
				},
				"code": {
					"type": "string",
					"example": "123445"
				}
			},
			"xml": {
				"name": "forgotPassword"
			}
		},
...
		"sendOtp": {
			"type": "object",
			"required": [
				"phoneNumber"
			],
			"properties": {
				"phoneNumber": {
					"type": "integer",
					"example": "7002708122"
				},
				"countryCode": {
					"type": "number",
					"example": "91"
				}
			},
			"xml": {
				"name": "verifyEmail"
			}
		},
		"verifyPhone": {
			"type": "object",
			"required": [
				"otp"
			],
			"properties": {
				"otp": {
					"type": "string",
					"example": "123456"
				}
			},
			"xml": {
				"name": "verifyPhone"
			}
		},
		"changePassword": {
			"type": "object",
			"required": [
				"oldPassword",
				"newPassword"
			],
			"properties": {
				"oldPassword": {
					"type": "string",
					"example": "123456"
				},
				"newPassword": {
					"type": "string",
					"example": "123456"
				}
			},
			"xml": {
				"name": "changePassword"
			}
		},
...
	}
}